Implement Eval mxj object, integrate timbre/Max.
Allows the evaluation of a preconfigured Clojure expression (in the
context of Afterglow) when a bang is received, or an arbitrary
expression sent as a string argument to an eval message.

Also ties the timbre logging system used by Afterglow to the Max
console, so that messages appear there, and errors are properly
indicated in red.